+++ nemo/collections/llm/recipes/deepseek_v3.py
@@ -123,7 +123,7 @@ def finetune_recipe(
             $ nemo llm finetune --factory "deepseek_v3(num_nodes=6, name='my_deepseek_v3_finetune')"
 
         Python API usage:
-            >>> recipe = finetune_recipe(name="deepseek_v3_finetune", num_nodes=2)
+            >>> recipe = finetune_recipe(name="deepseek_v3_finetune", num_nodes=6)
             >>> print(recipe)
 
     Note:
@@ -136,16 +136,18 @@ def finetune_recipe(
 
     if num_nodes is None:
         if peft_scheme is None or peft_scheme.lower() == 'none':
-            num_nodes = 64
+            num_nodes = 56
         elif peft_scheme.lower() in ['lora', 'dora']:
-            num_nodes = 6
+            num_nodes = 5
 
     recipe = default_finetune_recipe(model(), resume_path, dir, name, num_nodes, num_gpus_per_node, packed_sequence)
     if peft_scheme is None or peft_scheme.lower() == 'none':
-        recipe.trainer.strategy.expert_model_parallel_size = 64
-        recipe.trainer.strategy.pipeline_model_parallel_size = 8
-        recipe.model.config.num_layers_in_first_pipeline_stage = 7
-        recipe.model.config.num_layers_in_last_pipeline_stage = 6
+        recipe.trainer.strategy.sequence_parallel = True
+        recipe.trainer.strategy.expert_model_parallel_size = 4
+        recipe.trainer.strategy.tensor_model_parallel_size = 16
+        recipe.trainer.strategy.pipeline_model_parallel_size = 7
+        recipe.model.config.num_layers_in_first_pipeline_stage = 8
+        recipe.model.config.num_layers_in_last_pipeline_stage = 8
         recipe.optim.config.lr = 5e-6
     elif peft_scheme.lower() in ['lora', 'dora']:
         recipe.peft = run.Config(PEFT_STR2CLS[peft_scheme.lower()])
@@ -158,11 +160,12 @@ def finetune_recipe(
         ]
         recipe.optim.config.use_distributed_optimizer = False
         recipe.model.config.cross_entropy_loss_fusion = False
-        recipe.trainer.strategy.tensor_model_parallel_size = 1
-        recipe.trainer.strategy.expert_model_parallel_size = 16
-        recipe.trainer.strategy.pipeline_model_parallel_size = 3
-        recipe.model.config.num_layers_in_first_pipeline_stage = 20
-        recipe.model.config.num_layers_in_last_pipeline_stage = 20
+        recipe.trainer.strategy.sequence_parallel = True
+        recipe.trainer.strategy.tensor_model_parallel_size = 8
+        recipe.trainer.strategy.expert_model_parallel_size = 1
+        recipe.trainer.strategy.pipeline_model_parallel_size = 5
+        recipe.model.config.num_layers_in_first_pipeline_stage = 13
+        recipe.model.config.num_layers_in_last_pipeline_stage = 12
         recipe.optim.config.lr = 1e-4
     else:
         raise ValueError(f"Unrecognized peft scheme: {peft_scheme}")
